Lookup default AMIs.
If no default machine pool AMI is set, we will lookup the latest for
RHCOS and store it there for use on any pools that do not explicitly
store their own AMI. This lookup is done with code from
openshift-install.

In future we expect the RHCOS image to use to be a part of the release
image metadata.