Fix cluster destruction when install job deleted.
If a cluster were successfully installed, and the user manually deleted
the install job, code which updates the installed status was improperly
checking for a nil job, when it should actually have been checking for
an empty job struct. As a result deleting the install job would flip
cluster status to installed = false, triggering a fresh install job and
wiping out the existing cluster.