vendor: Bump openshift/installer to a7468d16b81d5fee95de4dc632fabe7b88d17f05
Generated with:

  $ emacs Gopkg.toml  # pin installer to a7468d16b81d
  $ dep ensure

using:

  $ dep version
  dep:
   version     : v0.5.0
   build date  :
   git hash    : 22125cf
   go version  : go1.10.3
   go compiler : gc
   platform    : linux/amd64
   features    : ImportDuringSolve=false

This pulls in the new pkg/destroy/aws (which is descended from our
previous contrib/pkg/awstagdeprovision).

I've also added non-go and unused-packages to the prune block, because
there's no need to ship those around when we don't need them.  For
example, this avoids pulling in the installer's newly-vendored
Terraform source.