Remove AMIs from API.
This feature was broken after we merged validation that you cannot
change anything outside of spec.Compute. Rather than adding another
exception, this moves the default AMI pinning to an annotation, as the
entire feature is not going to be needed once we implement the design
for OS image lifecycle, so the sooner it's out of the API the better.