Fix cluster version status on older cluster deployments.
Pre-existing clusters in opshive were erroring attempting to update the
annotations on the deployment due to very weird behavior with status
mandatory fields. (which shouldn't be involved as we're not updating
status)

Apply the fix for empty cluster version status fields, see if anything
changed, and if so save the status.