WIP: *: Extract installer from the target update payload
Take advantage of openshift/origin@59dab63d (Temporarily force the
installer to be part of the payload, 2018-12-08,
openshift/origin#21637) to avoid installer/update-payload mismatches.
This does not address install-config.yaml/installer mismatches, but
it's a step in the right direction.

Also:

* Replace 'shell' syntax highlighting with 'console', because these
  blocks have prompts.  And switch from the traditionally-PS2 '>' to
  the traditionally-PS1 '$' for those prompts.
* Stop exporting variables that are not needed by subprocesses.
* Drop OPENSHIFT_INSTALL_*.  This information is provided via
  install-config.yaml, and the installer ignores the environment
  variables since openshift/installer@6be4c253 (*: remove support for
  environment variables, 2018-12-10, openshift/installer#861).

There's an outstanding FIXME while I wait for guidance about who's job
it is to pull the installer image out of the update payload.  If we do
that in generate.go, we need to vendor some
not-really-designed-as-a-library origin code.  If we do it in the
launched container, we need both oc and Podman (or some other way to
actually run the discovered installer image).