remoteingress controller
extend ClusterDeployment to allow defining a list of ingress objects that should be configured on the cluster as IngressController objects for the openshift-ingress-operator to set up.

spec.ingress can be empty which will result in default cluster behavior (ie self-signed certs, *.apps.<base domain> default route, etc). otherwise, spec.ingress can provide for additional ingress objects (although you must provide settings for the 'default' ingress object in this case). and once you define ingress objects, you cannot go back to default behavior. all this is enforced by the admission webhooks.

when spec.ingress is defined, it will create a syncset object with all the desired ingress definitions so that the appropriate IngressController objects are set up in the cluster.

this version of the controller doesn't address having non-self-signed certificates created as part of the setup/sync process. this will come at a later date when the interaction between hive and the certificate request CRDs is finalized.