Remove generation of package bundle from OLM script.
We will limit this script to only generating a bundle for a specific
CSV. The packages file will be managed by the CD pipeline as it will
maintain a persistent store for the registry image mapping each version
to the one it replaces, for both stage/prod pipelines.

Adds a script to demonstrate how to generate your CSV and publish to a
registry, then subscribe.