Adds IMAGE_TAG
IMAGE_TAG is a convention of saasherder, the CI/CD tool used by the
AppSRE team:
https://github.com/openshiftio/saasherder/

This tool revolves around the idea of deploying OpenShift templates,
always passing `IMAGE_TAG` as a parameter generated by obtaining the
short git commit hash.

In this particular case we need IMAGE_TAG to appear in the OpenShift
template we are applying:
`/hack/olm-registry/olm-artifacts-template.yaml`.

Since we need to have more than one catalog stream, namely `production`
and `staging` we opted for the idea of generating:

- `staging-<hash>` and `production-<hash>` as the non-moving targets
- `staging-latest` and `production-latest` as the moving targets

When it comes to iterating the CatalogSource, we would need to oc apply
the non-moving target, so it kicks OLM and the new image catalog is
deployed.

It might seem a bit convoluted to have the image catalog `image`
parameter built this way:

```yaml
image: ${REGISTRY_IMG}:${CHANNEL}-${IMAGE_TAG}
```

but as far as I can tell, it's the only way we can combine `IMAGE_TAG`
and multiple streams (`staging` and `production`).

For more context, saasherder will apply this:
https://github.com/app-sre/saas-hive/blob/master/hive-services/hive.yaml

`REGISTRY_IMG` and `CHANNEL` are hardcoded, but `IMAGE_TAG` is generated
dynamically, based on the git commit.