Use remote machinesets for AMI lookup.
Installer code has changed since this was implemented, they now pin the
build in a shell script, compile it into the code as a constant (which
we can't see by vendoring), and we cannot specify the same build ID via
calling the function as it's a constant right now, not a parameter.

For now we will lookup the AMI to use by examining pre-existing machine
sets. This will need to be changed in future, but hopefully this will
get us by until we can extract the AMI from the release image properly.