Deliver a hive-admin role via the operator.
This role allows the basic cluster wide permissions we expect to need to
administer a hive cluster.

It grants read access to all the Hive types, as well as jobs, pods, and
pod logs.

It grants write access to the HiveConfig.

It does not grant access to secrets or configmaps.

This is delivered via the operator itself, rather than the CSV because
we need to bind to a group, not a service account.