Move possible 30s remote timeout into a guarded block.
Cluster deployment controller uses a remote client to lookup the default
route for setting the URL to the console in status. However this call
can timeout if the cluster has been destroyed externally.

Additionally we don't even need the client once the URL has been set one
time, which is mostly true, so this timeout is entirely unnecessary and
can likely always be avoided by moving it into the guarded block and
skipped if the URL is already set.