Fix SyncSet controller hotloop.
If a SyncSet resource was fundamentally broken and will never apply
(target namespace does not exist), the controller would update a
condition if the reason or message changed. The error message was
included in the message and happened to contain a temporary filename,
which is always changing. This would trigger a status update, a new
reconcile, and thus a hotloop.

Switch to using a static error message with a note for how we can
improve in future. Full details will still be in the logs.

This exposed that errors are not being returned from the controller at
all in these cases, so it would be several hours before we tried again.
This is also corrected to return an error to the controller-runtime
framework, which will let us re-try and go into exponential backoff
which tops out around 10 minutes by default.