Additional metrics and cleanup.
Add histogram for the number of restarts it took to install the cluster.
This is separate from the per-cluster metric we publish while
installing.

Revamp per cluster restarts metric and add a success bool 0/1 metric to
use in addition when alerting. (i.e. if restarts > 4 and success = 0 for
a cluster, alert) We stop reporting both once the cluster is
installed, though it will not be marked stale until Hive is restarted.

Add metrics for deprovisioning clusters and how long they've been in
that state, similar to uninstalled clusters. This will allow us to track
stuck cluster deletions.

Fix bugs in metrics when there are no clusters of a particular type.