Add provisioning underway seconds metrics.
Report a metric labelled by cluster name, namespace, and type, for any
cluster that is currently provisioning or deprovisioning. Value is the
number of seconds since it was created/deleted. Once the job completes
we will reset the value to 0 and then stop reporting it which should
limit cardinality.

This should be a better solution to the per-cluster metrics we want to
alert on. This change removes the two metrics we had prior which were
install restarts plus a success boolean metric, which had to be
correlated to know if there was still a problem.

This solution catches situations where no install job gets created, does
not require correlating any metrics against each other, and allows us to
easily alert after any time period we wish.