Delay install retries with sleep.
Modifies install manager to write a sleep-seconds.txt file that uses the
current number of retries for exponential backoff when an install fails,
up to a max of 24 hours. This will help prevent constant API calls on
clusters that need attention before they can possibly install
successfully.

Pod install command is now a shell script that optionally looks for this
file, and if found, sleeps the requested number of seconds. This should
allow the feature to be backward compatible with older hive images which
do not write the file.

This approach should limit the memory consumed as the hiveutil
install-manager has exited, and we're just left with a shell sleep.

Additionally this change adds a deprovison immediately after failed installs.

This will help reduce running resources in the event of a long delay to
the next attempt.

We will still attempt to cleanup at the start of a retry if the InfraID
is still set. It will not be if the deprovision after failure
succeessfully cleared it.