Fix leaking resources on install retries.
Recent change to cleanup immediately after failed installs had a bug
where we updated the cluster deployment with the new infra ID, but on a
different pointer to the object than we used later in the code to do the
cleanup, resulting in an attempt to cleanup an infra ID we already
cleaned up, and nothing is found, leaving cluster objects in place.

Issue surfaced as a DNS hosted zone already existing in this case.

Fix by setting the infra ID explicitly on the cd we use throughout the
install.