Monitor install logs for errors.
Adds a new controller to monitor the install logs we upload as
configmaps. Logs will be logged to the main hive-controllers log so we
can find them in one place.

Logs will be scanned for a list of regexes (defined in a configmap and
delivered by our operator) which map to keys. If encountered those keys
are reported as the reason on the InstallFailing condition, and a metric
is incremented.