WIP: save job spec hashes for install jobs
allow detection of an install job spec change by saving the hash of the job.Spec into an annotation on the job.

with this we can detect when we need to delete/replace an existing install job (for example when a new version of hive is installed and wants to launch install jobs with different configurations than the previous version (fixing a bug to get the install jobs unstuck)).