Copy installer binary to a temp file and rename.
Problems were spotted in a possibly heavily loaded stage cluster where
the installer container, which just copies a file, would fail with file
in use errors. We suspect the file was only partially copied when the
install manager saw that it exists and tried to execute it.

Instead we will copy to a temp filename, and then rename which is a much
quicker operation.