Attempt at fixing stale deprovision underway metrics.
We've seen these get stuck in prod, metric keeps reporting a value, but
the cluster is long gone, and the problem will not correct itself until
hive pod is restarted at which point we no longer report.

Cannot reproduce, or say for sure why, but it's possibly a race
condition between the two places this metric was managed. This change
moves it to one controller, and makes sure we clear it if we process a
deleted cluster deployment that doesn't have our finalizer.