Reduce configmap reconcile rate and logging.
This was reconciling on all configmaps and filtering down in the
reconcile loop, and logging a log entry for each. The configmaps used
for HA pod deployments are constantly updated leading to a steady stream
of reconciles and logging.

Reduce by filtering out any configmaps that do not look like hive
install log configmaps.