Skip reconciles for unreachable clusters.
This code mistakenly assumed that if you tried to hit a cluster that
already has unreachable condition set, we should return that as an
error. This however causes the controller-runtime framework to keep
retrying, with backoff, but still a lot of reconciling and errors
logged.

Instead we should check for the condition and skip reconcile, without
considering this an error.

In time, if the cluster comes back, the unreachable controller will
clear the condition at which point all these controllers will resume
normal operation.