Use latest hive image for processing image resolution job.
The log gathering PR merged earlier today missed the fact that image
resolution jobs use the new script, but an old hive image to process the
results of the script (if the ClusterImageSet predates the change). This
causes the CLI image to be resolved but never set on the
ClusterDeployment status. As a result the new code will not create an
install job until the CLI image is resolved and saved in status.

Temporary fix that will probably need to be cleaned up in this commit.
Uses the latest Hive image (i.e. what we're running the controllers
from, as set by the operator), falling back to the hive image in the
ClusterImageSet. This change is likely permanent considering the work
we're doing this week, only the fallback will likely be removed.