Force reloading of APIServer on certs change
Whenever the serving cert secret is updated (as in renewal), we need
to let the Kube API server on the target cluster know so that the
new certs are applied. This fix uses a hash of all cert secrets
passed to the API server to set the value of the
"forceDeploymentReason" field on the KubeAPIServer resource on the
cluster.