semver complains if the PATCH first char is a 0
OLM parses the version and uses github.com/blang/semver to validate it
as a semver.

We have found out that if the PATCH (third component) starts with
`0` it raises this error:

https://github.com/blang/semver/blob/1a9109f8c4a1d669a78442f567dfe8eedf982793/semver.go#L376

When this happens OLM refuses to deploy the CSV and no further updates
will happen.

After some testing, we decided to prepend the PATCH with the string
`sha` so this error is not raised.

```go
package main

import (
	"fmt"
	"os"

	"github.com/blang/semver"
)

func main() {
	v, err := semver.Parse(os.Args[1])
	if err != nil {
		panic(err)
	}

	fmt.Printf("%v\n", v)
}
```

```
$ ./testsemver 0.1.1036-0152295
panic: Numeric PreRelease version must not contain leading zeroes "0152295"

goroutine 1 [running]:
main.main()
	/home/jmelis/borrar/testsemver/main.go:13 +0x216

$ ./testsemver 0.1.1036-sha0152295
0.1.1036-sha0152295
```