Update makefile to run golangci lint bydefault
Also updated the installation method to make it work in CI systems.

Commented some of the check in golangci.yml because either
these need more effort to fix or not critical as compared
to making golangci-lint default in make test.

Once golangci-lint is enabled it will help fixing incoming
code for lint issues and we can gradually add more checks by
removing the comments in golangci.yml.

Signed-off-by: Lalatendu Mohanty <lmohanty@redhat.com>