Add ssh-agent to hive container for log gathering.
This was present in the Dockerfile.dev we were using, but not in the
final layer we based on for official builds. Unfortunately RHEL based
containers cannot run yum during builds due to lack of entitlements.
We're only using this layer to run precompiled binaries, so switching to
the universal base image which allows us to use yum freely.