allow for restoring clusterdeployment and clusterprovision
* The installed field in clusterdeployment has been moved from the status to the spec.
This allows the field to be retained when a clusterdeployment is restored. The installed
field in status is deprecated but will be retained until all existing clusterdeployments
are converted to use the installed field in spec.

* Adjust the validation for clusterprovision to accommodate adopting an existing
cluster. When adopting, the expectation is that the podSpec field will be omitted.
So when the podSpec field is empty, the validation ensures that the stage is "complete",
that the attempt number if 0, and that there is not a previous infra ID or cluster ID.

* If the clusterdeployment is marked as installed but there is not a reference to the
clusterprovision, then the controller will attempt to adopt an existing completed
clusterprovision. Upon adoption, the infra ID, cluster ID, admin kubeconfig secret,
and admin password secret will be copied from the clusterprovision to the clusterdeployment.