create dummy clusterprovision for legacy clusterdeployment
If a clusterdeployment was installed prior to the changes that brought
clusterprovisions to be, then the clusterdeployment will not have a
clusterprovision associated with it. This is not a problem for normal
operation. However, if the clusterdeployment is backed up and restored,
then the clusterdeployment will lose its infra ID and kubeconfig
secret reference, since those are stored in the status. Normally,
after a restore, these items--along with others--are copied back into
the clusterdeployment from the clusterprovision. So, in order for the
clusterdeployment to work after being restored, the controller needs
to create a dummy clusterprovision for the clusterdeployment with
the pertinent data to be restored.