installmanager: additional wait-for install-complete after failed install
If the `openshift-install create cluster` command fails after the bootstrapping
has completed, then run `openshift-install wait-for install-complete` once to
wait an additional 30 minutes for the cluster to finsih installing. It is not
uncommon for a cluster install to time out when the cluster is at 99+% complete.
This will give those clusters a chance to finish before tearing them down and
starting over.