controllers: dummy provision for clusterdeployment with installed timestamp
A dummy provision should be created for all legacy clusterdeployments. However,
a dummy provision is not being created for legacy clusterdeployments that have
an installed timestamp. To remedy this, the check for the installed timestamp
has been moved inside the if statement verifying that the clusterdeployment
has a link to a clusterprovision.

Additionally, the clusterdeployment controller has been changed to stop reconciling
a clusterdeployment that is installed, has no clusterprovision to adopt, and has
no infra ID. This is likely to happen when a clusterdeployment is restored prior to
its clusterprovision being restored. The clusterdeployment controller should under
no circumstances create a new clusterprovision when the spec of the clusterdeployment
indicates that the cluster is installed.