Skip minor OpenShift specific assets when running on vanilla Kube.
We deploy two ClusterRoles and ClusterRoleBindings intended to be mapped
to hive-admin and hive-reader groups in Hive clusters. Because these
groups only exist as OpenShift types, it appears using the OpenShift
variant of ClusterRole and ClusterRoleBinding is the only way to map to
them.

These two roles however are not required or useful for development, and
as such can be safely skipped.

hive-operator now checks if we're running on OpenShift by attempting to
list DeploymentConfigs. If this fails, these assets will not be
deployed and a warning is logged.

This change does not get the operator fully working on vanilla Kube,
next up will be fixed for hiveadmission certs.