controllers: do not compare hashes for merged pull secret
Compare the full pull secrets instead of hashes of the pull secrets.
Comparing the hashes is only helpful if we already have at least one
of the hashes. In this case, we only have the full strings of both
and are creating both hashes. It is more efficient, then, to simply
compare the full strings.