controllers: ensure all status update of clusterdeployment have valid clusterversion
The clusterdeployment controller makes an early attempt to fixup the clusterversion field
so that it passes validation. However, there are some other controllers that race with
the clusterdeployment controller to write to the status of a clusterdeployment. When one
of the other controllers wins the race, that update fails and emits an error log entry.
To avoid the race, the logic for fixing up the clusterversion field has been moved to a
wrapper around the status client. Consequently, every update of a clusterdeployment by
every controller will fixup the clusterversion prior to making the update.