Fix bootstrap log gathering on 4.2 by using openshift-install gather bootstrap.
For 4.1 we relied on terraform tfstate file to locate the bootstrap IP,
this file format has changed and we should not use an installer
implementation detail if possible.

In 4.2 openshift-install adds the "gather bootstrap" command to do this
work for us, without relying on any terraform internals.

This patch removes support for 4.1 bootstrap log gathering, we've never
seen it work in prod and 4.2 is out now. oc adm must-gather will still
be run against 4.1 clusters that make it past bootstrap, but still fail
to initialize.