controllers: ensure valid cd status for imageset jobs
The clusterdeployment status needs special massaging for it to be
acceptable to the validator. The controllers do this massaging
automatically, but the imageset job does not. To ensure that status
updates from the imageset job are accepted, these changes have the
clusterdeployment controller update the status prior to creating
the imageset job if the massaging has not yet been done.