Add InstallConfigSecret for v1 API.
To better support installer features as they land without requiring
exposure in Hive, we now accept an InstallConfig secret to be passed
directly to the installer. A secret was chosen as the InstallConfig
contains ssh key and pull secrets.

Some fields are now no longer required in the ClusterDeployment.

Updates create-cluster command and remotemachineset controller for this
change.